[Eir]: Phase 3 - Massive Testing - improved TXD and DFF handling (now more stable and faster) - added a new scripting function and two new events related to model streaming - fixed vehicle TXD not being registered by shaders - fixed a bug in RenderWare statistics collection
This update brings the stability of model handling from MTA:Eir to the Eir fork. From now on, custom models will not unexpectably fault anymore (neither their textures nor collision). Please test the custom model logic to its heart, and report bugs to the official MTA bugtracker (remember to flag it as Eir so I will find them).

New Scripting Stuff:
- bool engineIsModelBeingUsed( int modelIndex )
returns whether the model denoted by modelIndex is being used by MTA or not. Return false if there is not model info denoted by modelIndex.

- event onClientModelRequest( int modelIndex )
this event is triggered when MTA starts using a model info. The index of the model info is the first argument given to the callback. This event can be used to replace the model of an entity jsut before it streams in.

- event onClientModelFree( int modelIndex )
this event is triggered when MTA does not use a model info anymore. It is the counterpart to onClientModelRequest and could be used to free resources of a model info that are not required anymore. Note that these events are only used for when MTA entities request models, not GTA:SA buildings.

Have fun!
Any issues that need consulting? Visit me at irc.multitheftauto.com #mta.recore

+++ forks/eir/MTA10/game_sa/CTextureManagerSA.instance.cpp
@@ -98,7 +98,7 @@ static bool Txd_DeleteAll( RwTexture *tex, int )
 
 void CTxdInstanceSA::Deallocate( void )
 {
-    unsigned short id = TextureManager::GetTxdPool()->GetIndex( this );
+    int id = TextureManager::GetTxdPool()->GetIndex( this );
 
     // Notify the shader system
     OnStreamingRemoveTxd( id + 20000 );
@@ -192,7 +192,7 @@ void CTxdInstanceSA::InitParent( void )
     CTxdPool *txdPool = TextureManager::GetTxdPool();
 
     // Assign texture imports
-    unsigned short id = txdPool->GetIndex( this );
+    int id = txdPool->GetIndex( this );
 
     // Notify the shader system
     OnStreamingAddedTxd( id );
